requirements.txt

Pillow==10.3.0

requests==2.32.3

pyinstaller==6.3.0
